whattowatch/requirements.txt
asgiref==3.5.2
<<<<<<< HEAD
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
Django==3.2.6
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==21.3
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
selenium==4.6.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.4
wsproto==1.2.0
=======
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.6
django-allauth==0.50.0
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.12
>>>>>>> bf8c2e7808de8f109bd4d7ecd084d718e924e923
